feat!: remove `setup` function and initialize in `plugin` directory
